Rename 'sshproxy' package to 'proxy'.
